feat(deep_causality_data_structures): Removed all remaining unsafe code.
