+++ deep_causality_data_structures/Cargo.toml
@@ -40,10 +40,6 @@ path = "examples/window_type/vector_storage.rs"
 #name = "window_type_unsafe_vector_storage"
 #path = "examples/window_type/unsafe_vector_storage.rs"
 
-[features]
-default = []
-unsafe = [] # Enable unsafe implementations
-
 [dev-dependencies]
 criterion = { version = "0.7", features = ["html_reports"] }
 rand = { version = "0.9", features = ["small_rng"] }
