+++ deep_causality_data_structures/README_ArrayGrid.md
@@ -93,7 +93,7 @@ Because of the const generic, the get operation performs at a near constant time
 therefore the benchmark skips the get operation because there is nothing worth measuring.
 
 
-### Safe Implementation Performance Summary
+### Performance Summary
 - **1D Grid**: 604.71 ps (±5.01 ps) 
 - **2D Grid**: 581.33 ps (±2.81 ps) 
 - **3D Grid**: 862.16 ps (±13.60 ps) 
@@ -105,27 +105,10 @@ therefore the benchmark skips the get operation because there is nothing worth m
 2. Performance degrades progressively with higher dimensions, particularly noticeable in 3D and 4D operations
 3. The 4D grid operations show the most significant performance impact, taking nearly twice as long as lower-dimensional operations
 
-
-### Unsafe Implementation
-- 1D Grid: 543.38 ps 
-- 2D Grid: 414.32 ps
-- 3D Grid: 587.49 ps 
-- 4D Grid: 820.57 ps 
-
-## Safe vs. Unsafe Implementation 
-
-| Dimension | Safe Implementation | Unsafe Implementation | Improvement |
-|-----------|-------------------|---------------------|-------------|
-| 1D Grid   | 604.71 ps        | 271.38 ps          | 55.1%       |
-| 2D Grid   | 581.33 ps        | 417.39 ps          | 28.2%       |
-| 3D Grid   | 862.16 ps        | 577.04 ps          | 33.0%       |
-| 4D Grid   | 1.137 ns         | 812.62 ps          | 28.5%       |
-
 ## Technical Details
 - Sample size: 100 measurements per benchmark
 - Outliers properly detected and handled (2-8% outliers per benchmark)
 - All benchmarks were run with random access patterns to simulate real-world usage
-
 
 ## Hardware & OS
 - Architecture: ARM64 (Apple Silicon, M3 Max)
