+++ deep_causality_data_structures/benches/benchmarks/fields.rs
@@ -4,6 +4,5 @@
  */
 
 pub const SIZE: usize = 10;
-#[cfg(feature = "unsafe")]
-pub const CAPACITY: usize = MULT * SIZE;
+
 pub const MULT: usize = 100;
