+++ deep_causality_data_structures/benches/benchmarks/grid_type/mod.rs
@@ -4,6 +4,3 @@
  */
 
 pub mod bench_grid_array;
-
-#[cfg(feature = "unsafe")]
-pub mod bench_grid_array_unsafe;
