+++ deep_causality_data_structures/benches/benchmarks/window_type/bench_window_comp.rs
@@ -6,21 +6,12 @@
 use criterion::{BenchmarkId, Criterion, criterion_group};
 use std::hint::black_box;
 
-#[cfg(feature = "unsafe")]
-use deep_causality_data_structures::{
-    ArrayStorage, UnsafeArrayStorage, UnsafeVectorStorage, VectorStorage, WindowStorage,
-};
-
-#[cfg(not(feature = "unsafe"))]
 use deep_causality_data_structures::{ArrayStorage, VectorStorage, WindowStorage};
 
 const SIZE: usize = 4;
 const CAPACITY: usize = 1200;
 const MULT: usize = 300; // 300 * 4 = 1200 Same capacity for Vec impl as for Array impl
 
-//
-// Safe ArrayStorage
-//
 fn array_operations(c: &mut Criterion) {
     let mut group = c.benchmark_group("array_operations");
 
@@ -85,79 +76,6 @@ fn array_operations(c: &mut Criterion) {
 
     group.finish();
 }
-
-//
-// Unsafe ArrayStorage
-//
-#[cfg(feature = "unsafe")]
-fn unsafe_array_operations(c: &mut Criterion) {
-    let mut group = c.benchmark_group("unsafe_array_operations");
-
-    // Basic push operation
-    group.bench_function("push_single", |b| {
-        let mut storage = UnsafeArrayStorage::<i32, SIZE, CAPACITY>::new();
-        b.iter(|| {
-            storage.push(black_box(42));
-        });
-    });
-
-    // Push operation with rewind
-    group.bench_function("push_with_rewind", |b| {
-        let mut storage = UnsafeArrayStorage::<i32, SIZE, CAPACITY>::new();
-        for _ in 0..CAPACITY - 1 {
-            storage.push(42);
-        }
-        b.iter(|| {
-            storage.push(black_box(42));
-        });
-    });
-
-    // Sequential operations
-    group.bench_function("sequential_ops", |b| {
-        let mut storage = UnsafeArrayStorage::<i32, SIZE, CAPACITY>::new();
-        for i in 0..SIZE {
-            storage.push(i as i32);
-        }
-        b.iter(|| {
-            storage.push(black_box(42));
-            black_box(storage.first().unwrap());
-            black_box(storage.last().unwrap());
-            black_box(storage.get_slice());
-        });
-    });
-
-    // Batch operations
-    for size in [10, 50, 100].iter() {
-        group.bench_with_input(BenchmarkId::new("batch_push", size), size, |b, &size| {
-            let mut storage = UnsafeArrayStorage::<i32, SIZE, CAPACITY>::new();
-            b.iter(|| {
-                for i in 0..size {
-                    storage.push(black_box(i));
-                }
-            });
-        });
-    }
-
-    // Memory access patterns
-    group.bench_function("memory_access", |b| {
-        let mut storage = UnsafeArrayStorage::<i32, SIZE, CAPACITY>::new();
-        for i in 0..100 {
-            storage.push(i);
-        }
-        b.iter(|| {
-            for i in 0..10 {
-                storage.push(black_box(i));
-                black_box(storage.get_slice());
-            }
-        });
-    });
-
-    group.finish();
-}
-
-//
-// Safe VectorStorage
-//
 fn vector_operations(c: &mut Criterion) {
     let mut group = c.benchmark_group("vector_operations");
 
@@ -198,58 +116,6 @@ fn vector_operations(c: &mut Criterion) {
     group.finish();
 }
 
-//
-// Unsafe VectorStorage
-//
-#[cfg(feature = "unsafe")]
-fn unsafe_vector_operations(c: &mut Criterion) {
-    let mut group = c.benchmark_group("unsafe_vector_operations");
-
-    // Basic push operation
-    group.bench_function("push_single", |b| {
-        let mut storage = UnsafeVectorStorage::new(SIZE, MULT);
-        b.iter(|| {
-            storage.push(black_box(42));
-        });
-    });
-
-    // Sequential operations
-    group.bench_function("sequential_ops", |b| {
-        let mut storage = UnsafeVectorStorage::new(SIZE, MULT);
-        for i in 0..SIZE {
-            storage.push(i as i32);
-        }
-        b.iter(|| {
-            storage.push(black_box(42));
-            black_box(storage.first().unwrap());
-            black_box(storage.last().unwrap());
-            black_box(storage.get_slice());
-        });
-    });
-
-    // Batch operations
-    for size in [10, 50, 100].iter() {
-        group.bench_with_input(BenchmarkId::new("batch_push", size), size, |b, &size| {
-            let mut storage = UnsafeVectorStorage::new(SIZE, MULT);
-            b.iter(|| {
-                for i in 0..size {
-                    storage.push(black_box(i));
-                }
-            });
-        });
-    }
-
-    group.finish();
-}
-
-#[cfg(feature = "unsafe")]
-criterion_group! {
-    name = window_impl_comp;
-    config = Criterion::default().sample_size(100);
-    targets = array_operations, unsafe_array_operations, vector_operations, unsafe_vector_operations
-}
-
-#[cfg(not(feature = "unsafe"))]
 criterion_group! {
     name = window_impl_comp;
     config = Criterion::default().sample_size(100);
