+++ deep_causality_data_structures/benches/benchmarks/window_type/mod.rs
@@ -5,6 +5,5 @@
 
 pub mod bench_window_arr;
 pub mod bench_window_comp;
-pub mod bench_window_unsafe_arr;
-pub mod bench_window_unsafe_vec;
+
 pub mod bench_window_vec;
