+++ deep_causality_data_structures/src/lib.rs
@@ -27,8 +27,6 @@ pub use crate::window_type::SlidingWindow;
 pub use crate::window_type::storage::WindowStorage;
 pub use crate::window_type::storage_safe::storage_array::ArrayStorage;
 pub use crate::window_type::storage_safe::storage_vec::VectorStorage;
-#[cfg(feature = "unsafe")]
-pub use crate::window_type::storage_unsafe::{UnsafeArrayStorage, UnsafeVectorStorage};
 // num types
 pub use crate::num_types::one::*;
 pub use crate::num_types::zero::*;
