+++ deep_causality_data_structures/src/window_type/mod.rs
@@ -7,14 +7,8 @@ use std::marker::PhantomData;
 
 use crate::{ArrayStorage, VectorStorage, WindowStorage};
 
-#[cfg(feature = "unsafe")]
-use crate::window_type::storage_unsafe::{
-    unsafe_storage_array::UnsafeArrayStorage, unsafe_storage_vec::UnsafeVectorStorage,
-};
-
 pub(crate) mod storage;
 pub(crate) mod storage_safe;
-pub(crate) mod storage_unsafe;
 
 /// Returns a new sliding window with a vector storage and the size and capacity given as parameters.
 ///
@@ -54,64 +48,6 @@ where
     T: PartialEq + Copy + Default,
 {
     let storage = VectorStorage::new(size, multiple);
-    SlidingWindow::with_storage(storage)
-}
-
-/// Creates a new [`SlidingWindow`] with an unsafe vector-based storage implementation.
-///
-/// This function provides a more performant but unsafe alternative to the safe vector storage.
-/// It uses raw pointers internally for faster access but requires careful handling of memory.
-///
-/// # Feature Flag
-///
-/// This function requires the `unsafe` feature flag to be enabled:
-/// ```toml
-/// [dependencies]
-/// dcl_data_structures = { version = "0.4.7", features = ["unsafe"] }
-/// ```
-///
-/// # Arguments
-///
-/// * `size` - The fixed size of the sliding window
-/// * `multiple` - The growth factor for internal buffer allocation
-///
-/// # Type Parameters
-///
-/// * `T` - The type of elements stored in the window, must implement `PartialEq + Copy + Default`
-///
-/// # Returns
-///
-/// Returns a new [`SlidingWindow`] instance with unsafe vector storage
-///
-/// # Examples
-///
-/// ```rust
-/// # #[cfg(feature = "unsafe")]
-/// # {
-/// use deep_causality_data_structures::window_type::new_with_unsafe_vector_storage;
-///
-/// let mut window = new_with_unsafe_vector_storage::<f64>(5, 2);
-/// window.push(1.0);
-/// assert_eq!(window.first().unwrap(), 1.0);
-/// # }
-/// ```
-///
-/// # Safety
-///
-/// While the public interface is safe, the internal implementation uses unsafe code for
-/// performance optimization. The unsafe implementation maintains the following invariants:
-/// * Memory is properly allocated and deallocated
-/// * No out-of-bounds access occurs
-/// * No uninitialized memory is read
-#[cfg(feature = "unsafe")]
-pub fn new_with_unsafe_vector_storage<T>(
-    size: usize,
-    multiple: usize,
-) -> SlidingWindow<UnsafeVectorStorage<T>, T>
-where
-    T: PartialEq + Copy + Default,
-{
-    let storage = UnsafeVectorStorage::new(size, multiple);
     SlidingWindow::with_storage(storage)
 }
 
@@ -162,83 +98,6 @@ where
     [T; CAPACITY]: Sized,
 {
     let storage = ArrayStorage::new();
-    SlidingWindow::with_storage(storage)
-}
-
-/// Creates a new [`SlidingWindow`] with an unsafe array-based storage implementation.
-///
-/// This function provides a more performant but unsafe alternative to the safe array storage.
-/// It uses raw pointers internally for faster access and stack allocation for better cache locality.
-///
-/// # Feature Flag
-///
-/// This function requires the `unsafe` feature flag to be enabled:
-/// ```toml
-/// [dependencies]
-/// dcl_data_structures = { version = "0.4.7", features = ["unsafe"] }
-/// ```
-///
-/// # Type Parameters
-///
-/// * `T` - The type of elements stored in the window, must implement `PartialEq + Copy + Default`
-/// * `SIZE` - The fixed size of the sliding window, must be a const generic parameter
-/// * `CAPACITY` - The total capacity of the internal buffer, must be greater than or equal to `SIZE`
-///
-/// # Generic Constraints
-///
-/// * `T: PartialEq + Copy + Default` - Elements must be comparable, copyable, and have a default value
-/// * `[T; SIZE]: Sized` - The window size array must have a known size at compile time
-/// * `[T; CAPACITY]: Sized` - The capacity array must have a known size at compile time
-///
-/// # Returns
-///
-/// Returns a new [`SlidingWindow`] instance with unsafe array storage
-///
-/// # Examples
-///
-/// ```rust
-/// # #[cfg(feature = "unsafe")]
-/// # {
-/// use deep_causality_data_structures::window_type::new_with_unsafe_array_storage;
-///
-/// // Create a sliding window of size 3 with capacity 5
-/// let mut window = new_with_unsafe_array_storage::<f64, 3, 5>();
-/// assert!(window.empty());
-///
-/// // Fill the window
-/// window.push(1.0);
-/// window.push(2.0);
-/// window.push(3.0);
-///
-/// // Now we can safely check first and last elements
-/// assert!(window.first().is_ok());
-/// assert_eq!(window.first().unwrap(), 1.0);
-/// assert!(window.last().is_ok());
-/// assert_eq!(window.last().unwrap(), 3.0);
-/// # }
-/// ```
-///
-/// # Safety
-///
-/// While the public interface is safe, the internal implementation uses unsafe code for
-/// performance optimization. The unsafe implementation maintains the following invariants:
-/// * All array accesses are bounds-checked at compile time through const generics
-/// * No uninitialized memory is read
-/// * Memory is stack-allocated, providing better cache locality
-///
-/// # Compile-time Requirements
-///
-/// * `CAPACITY` must be greater than or equal to `SIZE`
-/// * Both `SIZE` and `CAPACITY` must be non-zero
-#[cfg(feature = "unsafe")]
-pub fn new_with_unsafe_array_storage<T, const SIZE: usize, const CAPACITY: usize>()
--> SlidingWindow<UnsafeArrayStorage<T, SIZE, CAPACITY>, T>
-where
-    T: PartialEq + Copy + Default,
-    [T; SIZE]: Sized,
-    [T; CAPACITY]: Sized,
-{
-    let storage = UnsafeArrayStorage::new();
     SlidingWindow::with_storage(storage)
 }
