+++ deep_causality_data_structures/tests/grid_type/mod.rs
@@ -7,8 +7,7 @@
 mod grid_array_tests;
 #[cfg(test)]
 mod grid_storage_tests;
-#[cfg(test)]
-mod grid_unsafe_test;
+
 #[cfg(test)]
 mod point_index_tests;
 #[cfg(test)]
