+++ deep_causality_data_structures/tests/mod.rs
@@ -3,4 +3,6 @@
  * Copyright (c) "2025" . The DeepCausality Authors and Contributors. All Rights Reserved.
  */
 mod causal_tensor_type;
+mod grid_type;
 mod num_types;
+mod window_type;
