+++ deep_causality_data_structures/tests/window_type/mod.rs
@@ -4,4 +4,3 @@
  */
 
 mod storage_safe;
-mod storage_unsafe;
