--- deep_causality_data_structures/tests/window_type/storage_unsafe/mod.rs
@@ -1,13 +1,0 @@
-/*
- * SPDX-License-Identifier: MIT
- * Copyright (c) "2025" . The DeepCausality Authors and Contributors. All Rights Reserved.
- */
-
-#[cfg(feature = "unsafe")]
-mod window_unsafe_array_stress_tests;
-#[cfg(feature = "unsafe")]
-mod window_unsafe_array_tests;
-#[cfg(feature = "unsafe")]
-mod window_unsafe_vector_tests;
-#[cfg(feature = "unsafe")]
-mod window_vector_stress_tests;
